en son hahahahhah

+++ app/src/main/java/com/app/chat/MessageModel.kt
@@ -5,11 +5,11 @@ class MessageModel {
     var byId:String?=null
     var message:String?=null
     var seen:Int?=null
-    var timestamp:Int?=null
+    var timestamp:Long?=null
     var toId:String?=null
     var id:String?=null
     constructor(){}
-    constructor(byId:String?,message:String?,seen:Int?,timestamp:Int?,toId:String?,id:String?){
+    constructor(byId:String?,message:String?,seen:Int?,timestamp:Long?,toId:String?,id:String?){
         this.byId = byId
         this.message = message
         this.seen = seen
